auth/requirements.txt
Flask
PyJWT
Flask-MySQLdb
Werkzeug
python-dotenv